Define the Uint8Array JSON envelope once in @workflow/world
world-local's fs.ts now re-exports the shared jsonReplacer/jsonReviver
instead of keeping a second copy of the same tagged-base64 envelope, and
deserializeQueueMessage views (rather than copies) Buffer inputs.
