+++ packages/world-local/src/fs.ts
@@ -1,7 +1,11 @@
 import { promises as fs } from 'node:fs';
 import path from 'node:path';
 import { EntityConflictError, WorkflowWorldError } from '@workflow/errors';
-import type { PaginatedResponse } from '@workflow/world';
+import {
+  jsonReplacer,
+  jsonReviver,
+  type PaginatedResponse,
+} from '@workflow/world';
 import { monotonicFactory } from 'ulid';
 import { z } from 'zod';
 import {
@@ -330,34 +334,9 @@ interface WriteOptions {
   overwrite?: boolean;
 }
 
-/**
- * Custom JSON replacer that encodes Uint8Array as base64 strings.
- * Format: { __type: 'Uint8Array', data: '<base64>' }
- */
-export function jsonReplacer(_key: string, value: unknown): unknown {
-  if (value instanceof Uint8Array) {
-    return {
-      __type: 'Uint8Array',
-      data: Buffer.from(value).toString('base64'),
-    };
-  }
-  return value;
-}
-
-/**
- * Custom JSON reviver that decodes base64 strings back to Uint8Array.
- */
-export function jsonReviver(_key: string, value: unknown): unknown {
-  if (
-    value !== null &&
-    typeof value === 'object' &&
-    (value as any).__type === 'Uint8Array' &&
-    typeof (value as any).data === 'string'
-  ) {
-    return new Uint8Array(Buffer.from((value as any).data, 'base64'));
-  }
-  return value;
-}
+// The Uint8Array-tagged JSON envelope is shared with the queue message codec
+// so the on-disk format and the queue wire format cannot drift apart.
+export { jsonReplacer, jsonReviver };
 
 export async function writeJSON(
   filePath: string,
